+++ include/Kmeans.hpp
@@ -118,13 +118,13 @@ public:
 class CoresetKmeans : public AbstractKmeans
 {
 private:
-    int mSampleSize;
+    size_t mSampleSize;
     std::unique_ptr<AbstractKmeans> pKmeans;
     std::unique_ptr<AbstractCoresetCreator> pCreator;
     std::unique_ptr<AbstractCoresetClusteringFinisher> pFinisher;
 
 public:
-    CoresetKmeans(const int& sampleSize, AbstractKmeans* kmeans, AbstractCoresetCreator* creator,
+    CoresetKmeans(const size_t& sampleSize, AbstractKmeans* kmeans, AbstractCoresetCreator* creator,
                   AbstractCoresetClusteringFinisher* finisher, IKmeansDataCreator* dataCreator,
                   std::shared_ptr<IDistanceFunctor> distanceFunc) :
         mSampleSize(sampleSize),
