+++ src/RandomSelector.cpp
@@ -18,7 +18,8 @@ int SingleWeightedRandomSelector::select(const std::vector<value_t>* const weigh
     return maxIdx - 1;
 }
 
-std::vector<int> MultiWeightedRandomSelector::select(const std::vector<value_t>* const weights, const int& sampleSize)
+std::vector<int> MultiWeightedRandomSelector::select(const std::vector<value_t>* const weights,
+                                                     const size_t& sampleSize)
 {
     std::vector<double> vals;
     for (auto& val : *weights)
@@ -35,7 +36,7 @@ std::vector<int> MultiWeightedRandomSelector::select(const std::vector<value_t>*
               [](std::pair<int, value_t> x, std::pair<int, value_t> y) { return x.second > y.second; });
 
     std::vector<int> samples;
-    for (int i = 0; i < sampleSize; i++)
+    for (size_t i = 0; i < sampleSize; i++)
     {
         samples.push_back(valsWithIndices[i].first);
     }
